64-bit conversion: disable game editor. Disable game editor for now because it stores pointers in RGuiItem::lId but lId needs to be (de)serialized as a 32-bit value.

+++ RSPiX/Src/ORANGE/GUI/guiItem.h
@@ -949,7 +949,7 @@ class RGuiItem : public RProps <U32, U32>
 																// when using transparent blit
 																// call.
 
-		intptr_t					m_lId;				// ID.  Used to identify this RGuiItem
+		int32_t					m_lId;				// ID.  Used to identify this RGuiItem
 														// from others.  See GetItemFromId().
 
 		Target				m_targetFocus;		// Target when focus received.
